[SPARK] Add common sql test utility methods to `DeltaSQLTestUtils` (#4131)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [X] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description
Adding the most commonly used methods in testing; such as getting the
delta log, snapshot, protocol, stats, columns etc in `DeltaSQLTestUtils`
so that we don't have to keep reimplementing them over and over again
and just use this trait instead.

## How was this patch tested?
No testing required.

## Does this PR introduce _any_ user-facing changes?
No.